Tanmay_Mudgal__Security_Simulator/Project/social_engineering_simulator/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=2.0.0
python-dotenv==1.0.0
scipy>=1.7.0
email-validator==2.1.0
Flask-Mail==0.9.1
google-generativeai>=0.3.0
stripe>=5.0.0
